aws-serverless-logistics/functions/pedidos/requirements.txt
boto3==1.34.0
PyJWT==2.8.0